first stab at pebble with the new iOS app

+++ core/main/m.h
@@ -1,3 +1,3 @@
 #pragma once
 
-typedef void (*message_callback_t) (const uint8_t* buffer, const uint16_t size, const uint64_t timestamp, const uint16_t duration);
+typedef void (*message_callback_t) (const uint8_t* buffer, const uint16_t size, const double timestamp, const uint16_t duration);
